pyproject.toml
[tool.poetry]
name = "shat-gpt"
version = "0.1.0"
description = "A short summary of the project"
authors = ["James Richardson <james.richardson.2556@gmail.com>"]
license = "Proprietary"
packages = [
    { include = "shat_gpt", from = "src" },
]
include = ["src/shat_gpt/res/*"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
importlib-metadata = {version = "^1.0", python = "<3.8"}
weaviate-client = "^3.20.1"
chainlit = "^0.4.0"
polars = "^0.18.3"
openai = "^0.27.8"
langchain = "^0.0.202"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^22.10"
pre-commit = "^2.20"
pytest = "^7.2"
pytest-cov = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
